backend/requirements.txt
fastapi
uvicorn
slowapi

playwright==1.49.0
sqlalchemy
psycopg2-binary
stripe
beautifulsoup4
python-multipart
requests
jinja2
openai
python-dotenv
pydantic-settings
python-jose[cryptography]
email-validator
pytest